refactor: use routes
